+++ learning/katas/go/core_transforms/additional_outputs/additional_outputs/task-info.yaml
@@ -17,5 +17,15 @@
 # under the License.
 #
 
-content:
-- Additional Outputs
+type: edu
+files:
+- name: test/task_test.go
+  visible: false
+- name: cmd/main.go
+  visible: true
+- name: pkg/task/task.go
+  visible: true
+  placeholders:
+  - offset: 969
+    length: 164
+    placeholder_text: TODO()
