+++ learning/katas/go/core_transforms/additional_outputs/additional_outputs/task-remote-info.yaml
