+++ learning/katas/go/core_transforms/additional_outputs/lesson-info.yaml
@@ -17,19 +17,5 @@
 # under the License.
 #
 
-type: edu
-files:
-- name: test/task_test.go
-  visible: false
-- name: go.mod
-  visible: false
-- name: cmd/main.go
-  visible: true
-- name: pkg/task/task.go
-  visible: true
-  placeholders:
-  - offset: 969
-    length: 164
-    placeholder_text: TODO()
-- name: go.sum
-  visible: false
+content:
+- additional_outputs
