+++ learning/katas/go/core_transforms/additional_outputs/lesson-remote-info.yaml
