+++ learning/katas/go/core_transforms/section-info.yaml
@@ -25,5 +25,5 @@ content:
 - flatten
 - partition
 - side_input
-- Additional Outputs
+- additional_outputs
 - Branching
